refactor: rename RadioButton to SimpleRadioButton
